refactor(profile): reorganize component positions within profile folder
