+++ src/app/app.module.ts
@@ -10,8 +10,8 @@ import { SearchBoxComponent } from './shared/components/search-box/search-box.co
 import { SharedModule } from './shared/shared.module';
 import { HeaderComponent } from './search-engine/presentation/home-page/components/header/header.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { ProfilePageModule } from './profile/presentation/profile-page.module';
-import { ProfilePageRoutingModule } from './profile/presentation/profile-page-routing.module';
+import { ProfilePageModule } from './profile/profile-page.module';
+import { ProfilePageRoutingModule } from './profile/profile-page-routing.module';
 import { LoginComponent } from './auth/presentation/login/login.component';
 import { RegisterComponent } from './auth/presentation/register/register.component';
 import { CardGroupComponent } from './profile/presentation/components/card-group/card-group.component';
