+++ src/app/profile/profile-page-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
-import { ProfileComponent } from './profile-page/pages/profile/profile.component';
+import { ProfileComponent } from './presentation/pages/profile-page/profile.component';
 import { RouterModule, Routes } from '@angular/router';
-import { ListGroupComponent } from './components/list-group/list-group.component';
-import { AboutmeComponent } from './components/aboutme/aboutme.component';
-import { NetworkComponent } from './components/network/network.component';
-import { ContactComponent } from './components/contact/contact.component';
-import { ArticleComponent } from './components/article/article.component';
-import { FingerprintComponent } from './components/fingerprint/fingerprint.component';
+import { ListGroupComponent } from './presentation/components/list-group/list-group.component';
+import { AboutmeComponent } from './presentation/components/aboutme/aboutme.component';
+import { NetworkComponent } from './presentation/components/network/network.component';
+import { ContactComponent } from './presentation/components/contact/contact.component';
+import { ArticleComponent } from './presentation/components/article/article.component';
+import { FingerprintComponent } from './presentation/components/fingerprint/fingerprint.component';
 
 const routes: Routes = [
   {
